Change to how target JLB is identified
John, don't merge this change yet, we're working a problem in Sydney. It seems the naming of the ELBs is sufficiently different for some stacks not to be identified by the GREP/CUT you were using. dig +short seems to be a suitable replacement but I'm making the change here so we can discuss it.